remove deprecated `Location.create`
