+++ packages/@ember/-internals/routing/index.ts
@@ -1,7 +1,6 @@
 // ES6TODO: Cleanup modules with side-effects below
 import './lib/ext/controller';
 
-export { default as Location } from './lib/location/api';
 export { default as NoneLocation } from './lib/location/none_location';
 export { default as HashLocation } from './lib/location/hash_location';
 export { default as HistoryLocation } from './lib/location/history_location';
