+++ packages/@ember/-internals/routing/lib/location/hash_location.ts
@@ -1,9 +1,7 @@
 import { set } from '@ember/-internals/metal';
 import { Object as EmberObject } from '@ember/-internals/runtime';
 import { bind } from '@ember/runloop';
-import { EmberLocation, UpdateCallback } from './api';
-import { getHash } from './util';
-
+import { EmberLocation, getHash, UpdateCallback } from './util';
 /**
 @module @ember/routing
 */
